requirements.txt
alembic==1.0.11
appdirs==1.4.3
attrs==19.1.0
bcrypt==3.1.7
black==19.3b0
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-API==1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.7
toml==0.10.0
Werkzeug==0.15.5
